Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This works fine, but it breaks translations.

Maybe we can come up with a way of having a large list of color names in CMParts which map led codes to what is shown in the UI. Then, the overlay only contains actual supported led codes which would be mapped to their names in runtime and translations could be done in  CMParts.

Patch-set: 1