Update patch set 1
Patch Set 1:

No problem. It jumps around when the fingers get to cllose to the same x (one on top of the other), but other than this seems to work quite well. It can actually be used by being careful to avoid getting there.

where is the code? I would not mind debugging it, but I'm still a bit lost in the maze.

Patch-set: 1