Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Not sure how is this different from http://review.cyanogenmod.com/#change,1802 , which I set up some time ago.

I have set this up on in froyo and a number of nightlies of gingerbread, and have never found any problem.

I guess it could be made into a configurable option if people is really scared by it, but it looks fairly innofensive. The behavior is not perfect but is also very usablel.

Patch-set: 1