Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

the last time i used it, it was more than unreliable. maybe you can live with it and have no problems, when maps turn right while you give the other input with your fingers - or everything flips in a way you didnt want to. but the normal user would see this as a bug and complain that we fix it. we cannot fix it - and when removing it, he will complain he wants it back.

the only way i personally can think of is a option in cmparts with a propper warning in the description

Patch-set: 1