Update patch set 1
Patch Set 1:

The index reported by the kernel is used to match to the index in LightSensor.cpp.  If these values are the same then everything works out because the kernel reported values are then matching the system light values.  If these values are different, then the system is using values that do not match what the kernel reported.  The user can observe that these are the correct indexed values in the custom light level settings.  As I said I have been using this on supersonic where it works as expected.  The only caveat is the top brightness level is never shown (I believe due to the light sensor kernel driver, you may have a better idea).

Patch-set: 1