Update patch set 1
Patch Set 1:

Right, the index is what matters, meaning if the index values in LightSensor.cpp are different from the index values that are being reported by the kernel, then the system will have the correct index, but different light values than expected.  This is how it works on supersonic.  If it works differently on bravo then I will leave it to you to figure out.  Out of curiosity, where else is this index read that it shouldn't be changed?  It has had no side effects here except that automatic brightness actually works.

Patch-set: 1