Update patch set 1
Patch Set 1:

You did also either run the script or update the device-vendor-blobs.mk manually, right? Otherwise it won't make a difference.

I did that on my N1 and it built fine but I have yet to notice much difference. I don't know if it even loaded the lib or not.

Patch-set: 1