Update patch set 1
Patch Set 1:

Tested this out my desire using last nights nightly build and the low battery led works correctly. That is to say once the battery is 10% or less the amber led flashes.

If you have a low battery AND a notification you get an alternating amber/orange led. This is to be expected as the the led is trying to notify you of two things at once (so my new code is invoked). Consider this a bonus feature of the patch, now notifications work when charging AND when in low battery state (previously notifications and low battery looked the same - amber flashing - but now you can tell the difference between having a flat phone and having a notification).

Note: it flashes with the same pattern as when charging and you have a notification. This is a shame but there are only so many pattern combos and colours available. And as for causing any confusion, if you see an alternating amber/green you are unlikely to be confused between charging+notification and flat+notification because in one scenario your phone will be plugged in and in the other it won't.

Anyway, to summarise: it seems to work fine. Low battery results in a flashing led warning.

Patch-set: 1