Update patch set 1
Patch Set 1:

I've noticed something like this with the Email app, sometimes when i open or want to reply an email once in a while the 'crashes' and put's me back to the email list instead of opening or reply/forward screen.

I think we'll need a logcat about when this happens but I don't have my desire anymore atm.

Patch-set: 1